pydantic to dataclass
