+++ .env
@@ -7,3 +7,4 @@ PROJECT_DIR=$(pwd)
 
 export PROJECT_DIR
 export PATH=$PROJECT_DIR/tools:$PATH
+export PYTHONPATH=$PROJECT_DIR/src
