+++ .github/workflows/checks.yml
@@ -16,6 +16,6 @@ jobs:
       - uses: ./.github/actions/python-poetry-env
         with:
           python-version: ${{ matrix.python-version }}
-      - run: poetry run tools/lint.sh
-      - run: poetry run tools/type_check.sh
-      - run: poetry run tools/test.sh
+      - run: poetry run scripts/lint.sh
+      - run: poetry run scripts/type_check.sh
+      - run: poetry run scripts/test.sh
