+++ pyproject.toml
@@ -28,10 +28,11 @@ packages = [
 
 [tool.poetry.dependencies]
 python = ">=3.9.1, <4.0"
-fastapi = "^0.85.0"
-uvicorn = "^0.18.3"
-requests = "^2.28.1"
+fastapi = "^0.115.6"
+uvicorn = "^0.34.0"
+requests = "^2.32.3"
 option = "^2.1.0"
+httpx = "^0.28.1"
 
 
 [tool.poetry.group.dev.dependencies]
