+++ scripts/lint.sh
