+++ scripts/type_check.sh
