+++ tests/scheduling_formatter/week_formatter_test.py
@@ -1,5 +1,5 @@
 from scheduling_formatter.week_formatter import \
-    format_week, format_from_formatted_days, WeekDays
+    format_week, format_from_formatted_days, WeekDays, WeekSchedulingPydantic
 from scheduling_formatter.day_formatter import Action, ActionType
 from utils.str_extensions import clean
 
@@ -43,3 +43,14 @@ def test_format_week():
 
     # Then r is the expected
     assert r == expected
+
+
+def test_to_week_scheduling():
+    # Given WeekSchedulingPydantic
+    scheduling_pydantic=WeekSchedulingPydantic(monday=[
+        Action(ActionType.open,3600),
+        Action(ActionType.close,7200),
+    ])
+
+    # And the expected WeekScheduling
+    scheduling
